Fix delete messages. Stuck on editing.

+++ frontend/components/client/messages/message_index_item.jsx
@@ -1,19 +1,49 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import EditMessageFormContainer from "./edit_message_form_container"
 
 
 class MessageIndexItem extends React.Component {
   constructor(props) {
     super(props);
 
+    this.state = {
+      editing: false
+    }
+  }
+
+  renderDeleteButton(){
+    if (this.props.currentUser != this.props.message.authorId) {
+      return null;
+    }
+    return (
+      <button onClick={() => this.props.deleteMessage(this.props.message.id)}>Delete</button>
+    );
+  }
+
+  openEditForm() {
+    this.setState({
+      editing: true
+    })
+  }
+
+  renderEditButton(){
+    if (this.props.currentUser != this.props.message.authorId) {
+      return null;
+    }
+    return (
+      <button onClick={() => this.openEditForm()}>Edit</button>
+    );
   }
 
   render() {
     return (
       <li>
-        {this.props.message.content}
-        <Link to={`/messages/${this.props.message.id}/edit`}>Edit</Link>
-        <button onClick={() => this.props.deleteMessage(this.props.message.id)}>Delete</button>
+        { 
+          this.props.editing ? <EditMessageFormContainer /> : this.props.message.content
+        }
+        {this.renderEditButton()}
+        {this.renderDeleteButton()}
       </li>
     );
   }
